compute-client: ensure controller commands are logged
This commit adjusts the `ActiveDurableState::method` to enforce a three
step process:

  1. Optionally validate the given controller command. Return an error
     if it fails to validate.
  2. Log the command to durable storage. Return an error if that fails.
  3. Apply the effects of the command to the system.

Step 2. is only mocked currently, at the moment the compute controller
still lacks access to durable storage.

Step 3. is not allowed to fail. The command has already been written to
durable storage and is therefore considered applied. We could maybe
attempt a rollback here, but that seems complicated. What if the process
crashes before the rollback can be completed? Then the restart code
would have to pick up and do the rollback, even though it cannot
differenciate between a failed command (that should be rolled back) and
storage corruption (that should probably lead to a panic).

There are a couple instances where applying the command to the system
(step 3) involves invoking the storage controller, which can fail. The
current code simply panics in this case. Assuming that storage errors
can only happen due to usage bugs (which we should fix) or transient
system failures (which should eventually resolve themselves after a
restart), panicking here should be fine.