test: Reset all system config before testdrive (#14576)
This commit adds a feature to testdrive that will reset all system
configurations at the start of each testdrive test. This is consistent
with our current approach of trying to reset Materialize's state at the
start of each test. This also helps protect against people forgetting
to reset configurations at the end of test, and configurations not
resetting due to failed tests.