test: Fix flaky storage test
The pg-snapshot-resumption test was flakey and sometimes failing. The
reason was that in step 03, Materialize was sometimes unable to select
a valid timestamp for the query and would hang forever.

The test purposely crashes STORAGED, which prevents the upper of t1
from advancing. However, the global timestamp does keep advancing. If
the global timestamp advances past the upper of t1, then we are unable
to select a valid timestamp under strict serializability, and the query
hangs forever.

Switching the isolation level to serializable for this test increases
the likelihood that we are able to select a valid timestamp for t1. Due
to #14696, there is still some possibility that the test will flake.
Once #14696 is fixed then this test should completely stop being flakey.

Fixes #14533