sql: back Date by i32
This matches postgres, and allows us to correctly constrain the domain
of the date datum. In addition to constraining to the document postgres
date range, we also constrain by the chrono NaiveDate range such that we
are guaranteed to be able to safely make that conversion. Many parts of
the code (avro encoding for example) have infallible methods that panic
instead of returning error. Using a constrained domain means those no
longer need to panic on invalid dates, and we still maintain postgres
compatibility.