Extract a compute controller
This commit extracts the compute controller code from the global
`Controller` and moves it into mz-compute-client. The intent of this
change is to introduce a clear boundary between the compute controller
and its clients, to make it possible to make that a boundary for durable
commands in a later step.