sql: Remove full and extended from SHOW commands
Most of the various SHOW statements support two modifiers: FULL and
EXTENDED. The FULL modifier typically adds additional columns to the
output of the SHOW statement, while the EXTENDED modifier typically
adds additional rows to the output of the SHOW statement with
information about system objects.

These modifiers are not well loved by users. See, for example, #14540.
The FULL modifier hides important information from users that ought to
be displayed by default. The EXTENDED modifier shows system objects,
but is redundant with showing objects FROM mz_catalog/pg_catalog.

We inherited these modifiers from MySQL in the very early days. Back
then, we were not all-in on emulating PostgreSQL; we tried to follow
MySQL when we inherited a statement from MySQL, and PostgreSQL when we
inherited a statement from PostgreSQL. So we tried to map MySQL's FULL
and EXTENDED modifiers to Materialize as best as we could, to
faithfully match MySQL.

Nowadays we're all in on PostgreSQL, and so we have the freedom to
break this loose compatibility with MySQL to redesign the SHOW commands
to our liking.

Fixes #14639