adapter: track constant-returning reads in explicit txns
Starting in #13382, we permit constant-returning reads in txns,
but in doing so ceased tracking if reads had occurred during a
txn. This commit refactors that behavior slightly to track that
a read occurred, albeit without a timestamp. I've added
additional logic to permit these reads during write-only txns.

This behavior is necessary to track if a query executed in an
explicit txns, which should prohibit it from returning to READ
WRITE mode.