storage: pg: handle excessive WAL lag edge-cases
When we consume the replication stream from a postgres database we only
get messages when there is data that is relevant to the publication that
we're interested in.

This can become problematic because receiving a message is the only way
we learn about progress statements, i.e the fact that up until a certain
LSN we have seen all the messages and we can therefore advance our
cursor.

One way this can happen is when the PostgreSQL instance contains
multiple databases and one of them is a high-traffic database but the
one we're streaming from is idle. In this case the WAL grows
continuously but we never advance our replication slot cursor,
eventually leading the instance to run out of disk space.

This PR fixes that by peeking into the replication slot changes on
initialization using a normal SQL query that invokes the
`pg_logical_slot_peek_binary_changes` administrative function. Since
this is a normal query we know when it has finished (as opposed to the
never ending replication stream) and we can observe if the result was
empty. If that's the case we know it's safe to fast-forward our WAL
cursor.

The source will continuously monitor replication lag and trigger a
replication restart if the lag grows beyond 100MB and we haven't
received any data message for more than 5 minutes.

Fixes #14694

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>