persist: fix bug where we'd incorrectly delete an initial rollup
This was causing maelstrom to flake in CI.

If two shard users race to initialize it, one will write a rollup and
only find out that it lost when attempting to CaS it in to the initial
consensus write. To avoid leaking this blob, we delete it when that
happens.

However, if we get an Indeterminate error in the CaS, but it actually
went through, then we'll "contend" with ourselves and get an expectation
mismatch. Previously, this would lead to us incorrectly deleting the
rollup blob, which manifests as a permanently hung shard. Instead, use
the actual fetched state to determine if our rollup actually made it in
and decide whether to delete based on that.

Closes #14575