coord: Make storage usage collection more precise
Previously, the Coordinator would collect storage metrics on a regular
interval using a tokio::time::interval. This was lacking in the
following ways:
  - Any time spent waiting would be forgotten in between restarts.
  - A crash loop could potentially cause repeated consecutive
  collections if we decided to blindly collect usages on restart.

This commit changes the logic to check when the previous collection was
and see how much more time needs to be spent waiting.

Fixes https://github.com/MaterializeInc/cloud/issues/3739