[sinks] Generated updated as_of in rehydrating client (#14785)
Previously, we always used the same as_of when the storaged instance was rehydrated. This meant that the following sequence would result in a panic:

CREATE SINK FROM SOURCE s (which has as_of t_source)
We create the sink with AS OF now() = t_sink > t_source
the persistent collection for source s is compacted so that t_source_1 > t_sink`
the storaged process crashes and is rehydrated
we recreate the sink using the cached command that uses t_sink
This panics when trying to read from the persist collection for the source s
We handled the case when the source was recreated on restart of environmentd by looking at the current since of the source in the storage controller. However, in this case, the storage controller doesn't instruct the client to restart -- the rehydrating task does. So we simply move the logic down a layer

Motivation
Fixes #14555