sql: rename mz_cluster_id to mz_environment_id (#14752)
* sql: rename mz_cluster_id to mz_environment_id

`mz_cluster_id` is a legacy function that returns a unique ID associated
with the Materialize deployment. We've since redefined "cluster" to mean
something else; the term that now means deployment is "environment". So
rename `mz_cluster_id` to `mz_environment_id`.

In the new cloud world, the environment ID is determined by the cloud
infrastructure, so it is provided via a new `--environment-id` command
line argument, rather than a UUID that is generated and stored in the
catalog on first boot. It defaults to a UUID of all zeros in tests and
local development.

* Reserve the old field name

Co-authored-by: Ben Kirwin <bkirwi@materialize.com>