sql: check type of referenced object in SHOW COLUMNS
`SHOW COLUMNS` should only work on object types which have columns. It
would previously silently return an empty result set for connections,
indexes, functions, secrets, types, and sinks; now it correctly reports
a usage error.

A future commit will make `SHOW COLUMNS` work with indexes, because
indexes are treated as having columns in PostgreSQL, and it makes sense
to be able to show the columns in an index.

Touches #14646.