replicated: report lower bounds of upper frontiers
This commit makes the `ActiveReplication` client report the lower bounds
of upper frontiers reported for each collection, in addition to the
upper bounds that were already reported before.
`ActiveReplicationResponse::FrontierUppers` is adjusted accordingly.

So far, the new lower bound is ignored by the compute controller. The
next commit will adjust the compute controller to ensure we don't allow
compaction beyond this lower bound.