Coord: Fix livelock in concurrent writes
If one writes tries to trigger a group commit, while another group
commit is already executing and holding the write lock, then the second
group commit will be deferred. A deferred group commit will wait and to
acquire the lock in an asynchronous task.

Previously, when the deferred group commit would acquire the lock, it
would immediately release the lock and immediately try and re-acquire
the lock. If the re-acquire failed, then it would defer itself again.
This would lead to a livelock situation where concurrent writes would
cause repeated defers. Not only would the writes hang, but the global
timestamp would be increased at an extremely rapid pace, eventually
putting it far into the future.

This commit changes the deferred group commit logic, so that once the
lock is acquired, it is not released until the group commit is complete.
This removes the live lock issue from concurrent writes.

Fixes #14793