sql: check all timestamps for a valid range
Previously we would allow any timestamp supported by chrono, even if
they weren't representable in pgwire/postgres or avro. Instead, wrap all
timestamps in a struct that verifies they are valid or returns an error.
Valid here means the smallest set of dates that are representable in
postgres, avro, and chrono.

This unfortunately hits a huge amount of code because many things
require infalliable casts from a rust type into a datum, but all of the
datetime logic works on chrono objects.