sql: Update output format of EXPLAIN TIMESTAMP
This commit updates the output format of EXPLAIN TIMESTAMP. It removes
the "has tables" and "table read ts" fields from the output because
they are no longer relevant in determining a timestamp. It also adds a
field to indicate if the query can respond immediately or if it needs
to block. This can help with debugging and help avoid misreading the
timestamps.