pg-cdc: Disable replica-identity-default-nothing.td
The test suffered from a race condition where the error that
the test was looking for would not happen unless the
replication is fully caught up. At the same time, there is
no easy way to instruct the test to wait until that happens.

Furthermore, the test was using a single replication stream
to test multiple things, and as a consequence was repeatedly
testing only the first of several expected errors. This needs
to be revisited again once Postgres sources switch away from the
"one-stream-per-postgres-database" approach.