adapter: rework sink progress topic exposure in system catalog
Now that mz_kafka_connections exists (#15013), we can restructure the
mz_sinks table to refer to the connection by ID, rather than directly
including the progress topic, for a more normalized schema. For
consistency, apply the same change to mz_sources.

Also rename the mz_kafka_connections.progress_topic column to
sink_progress_topic, to better reflect that it only applies to sinks
created on the connection.

Also rename the mz_kafka_sinks.sink_id column to id, for consistency
with the rest of the catalog.

Fix #14647.