catalog: Refactor catalog transactions
Catalog transactions take a list of operations. Previously,
transactions would iterate over that list twice. Usually in the first
pass, the catalog would update the stash transaction and generate
built-in updates. Usually in the second pass, the catalog would update
a copy of the in-memory catalog state. However certain operations do
not follow this pattern precisely.

This commit updates catalog transactions to combine both passes into a
single pass. This is useful for multiple reasons:
  * It is now possible to have some operation that depends on a
  previous operation in the same transaction. This can unlock things
  like executing multiple DDL statements in a transaction.
  * It is slightly easier to understand and reason about.
  * We can abstract all the logic to a method that doesn't take `self`.
  This makes it harder to accidently mutate `self` in the middle of a
  transaction.