coord: Update builtin fingerprint representation
This commit updates the representation of fingerprints for builtin
objects from a u64 to a String. Instead of using a hash of the builtin
object, we use a serialized JSON of the builtin object. This is because
hash collisions can cause us to mistakenly think that no migration has
occurred.

Additionally the builtin migration test was moved from a unit test to
an integration test. This is so it could have access to a test only
method.

Works towards resolving #11435
Follow up of #14369