compute-client,adapter: remove `Concrete` prefixes
`mz_compute_client` defines the `ConcreteComputeInstanceReplicaConfig`
type. This type, as well as its subtypes, have a `Concrete` prefix
solely because of a naming conflict with
`mz_sql::plan::ComputeInstanceReplicaConfig`. It is not great that the
compute client needs to concern itself with the type definitions in an
otherwise unrelated crate.

This commit removes the `Concrete` prefixes and instead solves the
naming conflict by prefixing the `mz_sql` type with its module name in
the only module where both are used.