adapter: introspection config in replica logging config
This commit adds the introspection config (log_logging and interval) to
the per-replica logging config, i.e. `ComputeInstanceReplicaLogging`.

The commit also includes a refactoring of the
`ComputeInstanceReplicaLogging` struct. It converts it from an enum into
a struct, to make dealing with it more convenient. Previously, the
`Default` enum variant was only needed as a placeholder for creating a
logging configuration for the default replica. This is now solved by
making the `sources` and `views` fields `Option`s in the
`SerializedComputeInstanceReplicaLogging` type.

The commit also renames various variables that hold
`ComputeInstanceReplicaLogging` instances from `persisted_logs` to
`logging`, to reflect that this struct contains general logging
configuration, not just IDs of persisted logs.