compute-client: use per-replica logging config
This commit changes the compute controller to make use of the logging
config passed to `add_replica_to_instance` instead of expecting an
instance-wide logging config.

When creating a new instance in the compute controller, we still have to
pass the IDs of the arranged log collection, since those are shared
between all replicas of an instance.