[sinks] More error handling in the Kafka sink (#14965)
* Panic on any persistent error

We only bail_err when the error is not retriable, or we have already
retried a bunch of times. When that happens it's safest to treat any
errors as unrecoverable.

* Error out if we're unable to fetch the progress timestamp

This one's a bit less obvious, since it seems pretty safe to retry, and
we're right at the beginning of the block. But I said I'd start with the
agressive version, so let's go full tilt.