adapter: simplify rules for unstable catalog items
Simply declare that all items in the mz_internal schema are unstable
while all items in the mz_catalog schema are stable. This seems a bit
easier to reason about.

It also unlocks building stable interfaces on top of unstable
interfaces. We can add views to mz_catalog that provide stable views on
top of unstable tables in mz_internal. This will be particularly useful
for billing: we want the flexibility to evolve the shard ID <-> object
ID mapping over time, but we're willing to expose a stable interface to
storage usage by object ID today.