testdrive: clean up kafka state at beginning of test script
The testdrive philosophy is to leave around state at the end of the
execution. Instead cleanup of the *last* test script is performed at the
beginning of each test run.  This makes it easy to debug a testdrive
script, because the state is left around after testdrive exits.  It also
makes it possible to use a testdrive script in demos or development as
to set up a Materialize instance for further tinkering.

The one wrinkle with this approach is that Kafka is not always
available. So condition the cleanup of Kafka state on whether the test
script uses a Kafka-related action.