repr: replace strconv-based parse_duration with humantime crate
The `humantime` crate provides a very lightweight `parse_duration`
function that accepts a syntax that works well in CLI arguments. This
commit uses that function consistently across the various binaries that
need to accept durations as command-line parameters.

The old approach was to use our built-in SQL cast for converting a
string to an interval. This worked well for e.g. `environmentd`, which
pulled in the SQL crates already, but was undesirable for the test
binaries in the `persist` crate, which did not want to pull the SQL
crates just to parse durations in its CLIs.