compute controller: don't crash on hydration
Discovered in #15132

When a replica crashes during hydration, environmentd also crashes
because it expects the command_tx channel to work. This PR changes warns
+ swallows the error instead. This (should be/) is correct, as the only
case where the command_tx send error happens is if the replica_task in
environmentd terminates. Thus the replica client is gone forever and we
dont risk to run into a partially hydrated replica. On the next
send/recv the error will cause a rehydration. We do not attempt to
eagerly rehydrate to ensure we don't fall into an infinite loop during
the rehydration.

This PR does not add any logic for dealing with a replica that reproducible
crashes on a received command.