compute: cease emitting FrontierUppers for subscribes
When a replica receives an `AllowCompaction([])` command, it emits a
`FrontierUppers([])` in response, to tell clients that only observe
responses that they can drop tracking the respective collection. This
logic had a bug where the replica would also emit those final
`FrontierUppers` responses for subscribes. This is incorrect because
subscribe progress is not communicated through `FrontierUppers` but
through `SubscribeResponse::Batch`es instead.