adapter: notify on replica changes during peeks
When a replica state changes, send a pgwire NOTICE messages informing
the users of the state change. This is useful in observing that replicas
are OOMing or in a restart loop which sometimes presents as a slow
query. The slow query gives us an opportunity to say something to a psql
console that will hopefully make it easier for users to figure out how
to fix the slow query problem.

Implement by adding a new notices channel on Session, and handing off a
send half to the Coordinator. We already send pending notices at various
times. Teach COPY, SELECT, and SUBSCRIBE to be able to also listen for
notices while waiting for rows.

See #15196