environmentd: serve all external routes on internal HTTP server too
Make the internal HTTP server a strict superset of the external HTTP
server. This makes it possible to e.g. use the memory visualizer even if
you only have access to the internal HTTP server.

Also move the jemalloc CPU/memory profiler to the internal SQL port.  It
exposes deep implementation details that aren't relevant to users, and
is likely just to be confusing, since it only displays profiles
environmentd', not computed or storaged.