storage: heartbeat write handles in (compute) persist_sink
Before, we were heartbeating the "writer" write handle, but only in
response to new updates or the frontier advancing.

Now, we heartbeat both the "writer" and "appender" write handles and we
make sure to do so independently of changes in operator inputs.

This fixes a bug that we observed in production where it can take a long
time to read the initial snapshot of the sinked collection, leading to
our write handle lease being expired.

I manually checked that this usage pattern of `async` timely operators
does not cause us to go into a busy loop.

Fixes #15191