adapter,expr: reject unmaterializable functions in WHERE clauses
The logic in `prep_relation_expr` was not being careful to look for
unmaterializable expressions in non-temporal expressions in an MfpPlan,
which could allow invalid references to unmaterializable functions in
WHERE clauses.

Also improve the error message for current_timestamp specifically to
give a hint to use mz_now instead.

Fix #15171.

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>