adapter: don't allow subscribing to arranged log sources
When a SELECT references log sources, we are careful to check those
references for validity. We require that a specific replica be targeted,
if the cluster has multiple replicas, and we require that targeted
replica has introspection enabled.

We were not performing the same checks for SUBSCRIBEs. A SUBSCRIBE that
referenced a log source in a cluster with introspection disabled would
cause the cluster to enter a permanent crash loop, since the log source
did not actually exist.

So this commit entirely removes the ability to SUBSCRIBE to arranged log
sources. We could in theory allow SUBSCRIBEs to target a specific
replica, but writing this logic is more difficult than just disabling
them entirely.