testdrive: generalize the kafka-verify-commit action
Change the kafka-verify-commit action to take the expected consumer
group ID directly, rather than having the action hardcode the rules for
converting a source name into a consumer group ID. This increases the
action's flexibility. Use the new flexibility to additionally test the
behavior of the GROUP ID PREFIX option.