switch reclocking data format to rows
this PR is the minimum surgery required in the current code that
preserves all the reclocking logic as-is but changes the remap shard to
be identically typed to the one of storage collections.

Reclock bindings for kafka sources are represented with two columns. An
int32 for the partition id and a uin64 for the offset.

Reclock bindings for postgres sources are represented with just one
column for the offset.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>