doc/user: polish v0.27.0 release notes
Flesh out the v0.27.0 release notes. They look plausible now. Surely
there are still some items that are missing, but this covers the
highlights.

Also update some of the text around release mechanics (e.g.,
the backwards compatibility policy and the new meaning of alpha) to
reflect the new reality.