storage: use wall-clock time to drive forward ingestion timestamps
This, together with the next commit makes the ingestion pipeline work
with multiple timely workers again.

The problem was that the batch counter that we use to drive timestamps
in the pipeline is not correlated between different workers, so a "slow"
worker would block downstream progress for the "fast" workers.