storage: in sources, track max-seen-offset and offset frontier separately
This, together with the previous commit that uses wall-clock time to
drive timestamps within the ingestion pipeline fixes a subtle bug.
Together, these two also fix the pipeline to work with multi-worker
setups again.

Before a recent change that made the ingestino pipeline multi-worker
aware, `ReclockOperator::reclock()` would always succeed, and it would
mint a binding for `max_seen_offset + 1` (that is, the upper implied
by the batch of updates that we’re seeing). The `source_upper` , which
was updated to either `offset` or `offset + 1`, depending on if the
message is a “final” message or not, was only used to drive forward the
reclocked frontier.

After this recent change, but before the changes in this commit,
`source_upper` was sent to `remap_operator`, which uses `mint()` to mint
new bindings. It took the upper as is, there was no +1 applied. The
`reclock_operator` was at the mercy of `remap_operator` to provide it
with the bindings it needs for reclocking batches of messages.
`source_upper` was also used to drive forward the frontier that in the
end will drive forward the reclock upper.

The solution, implemented in this change:
 - we track the `source_upper` and `max_seen_offset` separately (before,
   max seen timestamp was implied by the batch of messages that was
   being reclocked)
 - we send `max_seen_offset` to the `remap_operator`, we translate that
   to a "frontier" by adding `1`, which basically restores the old
   behavior and we don’t have to wait before we can reclock batches of
   messages
 - we use `source_upper` purely to track the upper, meaning we update it
   correctly based on `is_final` or not
 - we get rid of most of the complicated `is_final` logic in the source
   reader operator. We needed it because we tried to shoehorn
   `source_upper` into this dual use

Fixes #15300 (together with previous commit)