storage: in ingest pipeline, fix processing of updates for source_upper
Before, the last update (for the `global_source_upper`) would not be
processed because we only checked whether to apply an update when
receiving new updates. This meant that multi-worker sources where not
advancing their output/reclocked frontier when some source workers shut
down. This is the case for postgres sources, for example, which would
emit only one `SourceReaderOperatorOutput` and then shut down.

Plus, there was another subtlety: there was a block that would process
any remaining updates if the frontier goes to the empty frontier. This
only happens when all upstream source reader operators finish, though,
and not when only one of them finish.

Now, we make processing of these updates more "timely". We stash them as
they come in and process them once they are not beyond the frontier (tm)
anymore.