adapter: set compute read policies for persisted logs
This commit makes adapter initialize compute read policies for the
persisted log collections. Previously, only storage read policies would
be initialized for those, which meant that in the compute controller
persisted log collections would always have a read frontier of 0, and
consequently, no `AllowCompaction` commands would be sent for these
collections.

This didn't lead to issues because the storage controller is responsible
for managing the read frontier of storage collections. But for
consistency with other persist sinks in compute (materialized views), we
should still track the read frontier of log collections correctly in the
compute controller.