cloudtest: Use AWS S3 instead of Minio for S3 testdrive tests
Unfortuniately Minio does not seem compatible with S3 sources,
so pass through AWS credentials from the host to testdrive if
S3 testdrive tests are to be run.

The AWS credentials are generated by the scratch-aws-access Buikdkite
plugin in the form of environment variables that we copy over into
the testdrive container.

An --aws-region option can be passed to pytest which in turn
passes it to testdrive.

To avoid conflicts between concurrent tests that use S3, an actual
random seed must be used by the Platform Checks in a cloudtest environment.