persist: call maybe_heartbeat_reader in next_listen_batch loop
There might be some holdup in the next batch being produced. Perhaps
we've quiesced a table or maybe a dataflow is taking a long time to
start up because it has to read a lot of data. Heartbeat ourself so we
don't accidentally lose our lease while we wait for things to resume.

We think this was the cause of some expired reader panics seen in
production.

Closes #15396