Disable persisted introspection sources
This commit disables the creation, and consequently maintenance, of the
persisted compute introspection sources by default. They can still be
switched on through the `--persisted-introspection` command line flag.

This is meant to be a temporary measure to give persist some room to
breathe until the current compaction issues are resolved.