sql: don't ever interpret idents as objects for string options
This is a partial reversion of c7a4c0ac2. The intent was to avoid
a special case in the parser for options that take objects as values;
unfortunately this introduced *more* special cases than it saved. In
particular, it introduced a special case that leaked out into the user
experience: idents that named valid objects in the current schema could
not be coerced to strings, while idents that named invalid objects
could.

Seems better to just special-case the options that take values as
objects in the parser, and avoid the special cases everywhere else.

Follow up from #15113.