storage: in reclock.rs, make time control more explicit in tests
Before, we were using `tokio::time::Instant::now()`, which allows
advancing time via `tokio::time::advance()` but also automatically
advances the time if there is now work to do AND there is a
`tokio::time::sleep()` that is waiting.

The introduction of the new heartbeat tasks (for read/write handles), in
a separate PR, would futz with this time, because its
`tokio::time::sleep()` invocations can advance the time at
indeterministic points, messing with the tests.

Now, we control time explicitly using a fully custom `NowFn`. This also
changes the tests to be explicit about how the persist "cache" is
shared.