coord: Normalize storage update timestamps
Previously the timestamp of a row in the storage_usage table
corresponded to when that row was inserted. This resulted in many rows
that were inserted as part of the same update having timestamps that
differed by a couple of milliseconds, making it hard to aggregate the
rows from a single update.

This commit fixes this issue so that all rows from the same update
contain the same timestamp.

Fixes #15464