Add is_empty to MergeBatch and SpineBatch
Add a `is_empty` function to both MergeBatch and SpineBatch types, which
is more idiomatic than comparing the length to zero. It might have a
performance benefit because it's short-circuiting when calculating that
all parts are empty, which the length-based check does not do.

I haven't tested whether this actually improves performance, but it
seems to be a net-positive change.

This change implements a not previously described feature.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>