sql-parser: support option value sequences with parentheses
We use array-valued options in two user-facing places in our syntax:

    CREATE CONNECTION ... TO KAFKA (BROKERS ['a', 'b'])
    CREATE SOURCE ... FROM KAFKA CONNECTION ... (START OFFSET [0, 1])

This matches how you represent an array in SQL (`ARRAY[1, 2]`), but is
inconsistent with how DDL statements typically represent sequences of
things, e.g.:

    CREATE INDEX ... ON t (col1, col2)

This commit allows array-valued options to delimit their options with
*either* square brackets or parentheses. The following syntax
immediately becomes valid:

    CREATE CONNECTION ... TO KAFKA (BROKERS ('a', 'b'))
    CREATE SOURCE ... FROM KAFKA CONNECTION ... (START OFFSET (0, 1))

The primary impetus is to allow consistent extension of the `BROKERS`
option with the forthcoming AWS PrivateLink connection, which proposes
to use the syntax:

    CREATE CONNECTION ... TO KAFKA (
        BROKERS (
            'a' USING AWS PRIVATELINK s1,
            'b' USING AWS PRIVATELINK s2 (PORT 9093)
        )
    )

By allowing parentheses instead of square brackets, it's a bit clearer
that it's the same underlying `BROKERS` option, just with an optional
`USING AWS PRIVATELINK` clause attached to the elements of the sequence.

We cannot entirely remove the square bracket syntax, as that would be a
backwards incompatible change. We do, however, manage to eliminate the
`Value::Array` enum variant internally, which cleans up a match
statement in the code.