storage: in remap_operator, restart ReclockOperator from right before resume upper
With the band-aid, we allow compaction only up to around 3 days before
the resumption frontier. We still restart from right before the upper
which is a) perfectly safe to do, and b) will make extra certain that
the `as_of` is valid with regards to the since. If we pick the same time
that we compact up to, we might run into the same (alleged/potential)
off-by-one bug that this three-day-bandaid below is trying to paper
over.