storage: in ReclockOperator, use shard since instead of given as_of to seed self.since
My previous commit changes resumption to use a different frontier for
compaction and what we use as `as_of` on restart. Without this change,
we would initialize `self.since` to `resume_upper - 1`, but then later
try to `compact()` it to `roughly_resume_upper - 3 days`, which would
fail. It seems more straightfoward to just us the shard `since` for
initialization. We now use the `as_of` only to assert that it is valid,
given the shard `since`, that is, that we can serve the bindings that we
need to "serve".