compute-client: maintain peek read holds long enough
This commit changes the compute controller to defer dropping of read
holds installed for peeks until all replicas have responded to the peek
in any way. This prevents the controller from dropping the read hold too
early, namely when the fastest replica has responded to a peek but slow
ones might still be reading from the source collections at times that
could be compacted away after dropping the read hold.