timely-util: efficient type to represent partitioned timestamps
This PR intoduces a timestamp type that is able to represent a
partitioned timestamp domain, e.g the one of multiple independent Kafka
partitions, in a way that leads to efficient frontier representations
using `Antichain`s. This in turn enables writing timely operators that
work on data streams timestamped in such partitioned time domains.

STORAGE currently does a lot of work to extract `(PartitionId,
MzOffset)` timestamped messages from a `SourceReader` and reclock them
to the `mz_repr::Timestamp` domain internally to the opreator before
releasing them for further processing because otherwise the timestamps
are un-representable. This also forces us to put the original timesatmp
of messages as part of the `SourceMessage` that is being sent around.

Adopting this timestamp in a subscope of the ingestion dataflow graph
would allow us ship the data to downstream operators as-is and keep
working on it while the reclock operator is busy writing bindings to
persist. Then at the very last moment we can reclock the source
timestamp into `mz_repr::Timestamp` and write it to persist, having
saved some time in the meantime.

Additionally, it will allow us to express the reclocking operation as a
general transformation from *any* timely timestamp without requiring
specialized implementations like `OffsetAntichain`.

Finally, if we do build those abstractions we can take it one step
further and allow each source implementation to declare the timely
timestamp it wants to output data with, and as long as it is a valid
timely timestamp, we can reclock it. This would allow sources that don't
have the concept of partitions (like postgres) to simply output `u64`
timestamped data.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>