adapter: Add user permissions framework
This commit adds a framework for specifying and enforcing user level
permissions. The permissions indicate what clusters and schemas a user
is allowed to interact with as well as what actions that user is
allowed to perform. The permissions check is done right after statement
planning, but before sequencing.

The mz_introspection user is given a set of permissions to restrict
what it's allowed to do.

The permissions framework is non-internal users to prevent users from
interacting with the mz_system cluster, which was previously done in an
incomplete way.

Closes #15520