replicated: separate tracking of served peeks
This commit refactors the `ActiveReplication` client to separate the
tracking of unserved peeks from the tracking of served peeks by giving
each category its own field in the struct.

This refactor makes obvious bug we had previously, namely that when a
new replica was added, we would rehydrate it with peek commands that we
already served, needlessly wasting resources. This is fixed in this
commit as well, by rehydrating new replicas only with the unserved
peeks.