compute-client: remove the ActiveReplication protocol
This commit purges any traces of the old protocol between the compute
controller and `ActiveReplication`. Instead of producing
`ActiveReplicationResponses`, the `ComputeResponse` handlers now apply
the respective effects to the compute controller's state directly.

Similarly, removing a replica now directly triggers the appropriate
frontier updates to the compute collections, instead of emitting an
`ActiveReplicationResponse::FrontierUppers`. This introduces an async
method call to the `rehydrate_replica` method, rendering it
probably-not-cancel-safe. For this reason, this commit also removes the
`rehydrate_replica` calls from the `send` and `recv` methods and instead
arranges for rehydration to be performed inside `process`, which is not
required to be cancel-safe.