compute-client: remove ActiveReplication
This commit removes the `ActiveReplication` client by merging its
functionality into the compute controller. The intention is to keep the
diff as small as possible by mostly copying code around and avoiding
additional cleanups. In particular, the command-response protocol that
existed between the compute controller and `ActiveReplication` is still
used within the compute controller, even though it is a needless
indirection now. Subsequent commits will clean this up.