persist: stop recomputing since in expire_reader (for now)
Temporarily disabling this because we think it might be the cause of the
remap since bug. Specifically, a storaged process has a ReadHandle for
maintaining the once and one inside a Listen. If we crash and stay down
for longer than the read lease duration, it's possible that an expiry of
them both in quick succession jumps the since forward to the Listen one.

This is okay because since is always allowed to be behind (we take
advantage of this in maybe_downgrade_since, for example). It will jump
ahead on the next explicit downgrade_since call.