storage: in reclock_operator, make capability downgrading more lenient
The `remap_operator` will only hand out capabilities (with remap trace
updates) for timestamps that are beyond the reclocked frontier. When
this here `reclock_operator` is starting up, it can happen that the
frontier (as far as we are aware of) is still at `[0]`, meaning we
cannot downgrade our capabilities. This is perfectly fine, though, and
we will eventually downgrade the capabilities that we have when the
reclocked frontier advances sufficiently.

I also added some more trace logging that should make it easier to debug
future issues, if any.

Fixes #15558