sql: improve int + uint interoperability
Previously, there was no best common type between ints and uints
of the same width and functions taking both as arguments would
often default to choosing the float64-based implementation of a
function.

This commit improves both of those situations by teaching the
type system that using a signed int wider than the uint is
acceptable.