test: Fix test flake on pgwire::test_pgtest
The `pgwire::test_pgtest` would occasionally fail in CI with an error
that said:

  thread 'test_pgtest' panicked at 'called `Result::unwrap()` on an
  `Err` value: Interrupted system call (os error 4): waiting for
  ReadyForQuery, saw []

The error was coming from a tcp network read call.
https://doc.rust-lang.org/std/io/trait.Read.html#tymethod.read
According to the documentation, this call can fail with a non-fatal
error of the `ErrorKind::Interrupted` kind. In this scenario we should
be retrying the read.

This commit adds a retry when we see this error instead of failing the
test.

Fixes #15419