storage: in rehydrating client, be lenient about extra upper updates
It can happen during source shutdown that we remove the tracked upper
from our state but a `FrontierUppers` response is still on the wire.

This is very fine to ignore, especially now that these upper updates are
plain `Antichains`, and not `ChangeBatches` where we need to be extra
careful about not messing up our state.

Fixes #14746