stash: Add batched APIs for update and remove
The APIs for updating and removing items in a stash transaction only
accepted a single item at a time and had a runtime that was linear with
respect to the amount of items in the stash. These methods were
sometimes called in a loop, resulting in O(n^2) runtimes. This was one
of the causes for slow startups when there was a lot of items in the
catalog.

This commit adds batched APIs for update and remove so multiple items
can be removed or updated at once. This removes the O(n^2) runtimes and
greatly speeds up startup.

Works towards resolving #15380