stash: use single-statement consolidate
Previously consolidation happened by fetching data into rust,
consolidating, the inserting the rows back. This runs one INSERT
statement per unique (k,v) pair, which is quite a lot of rows moving
back and forth. Change this to consolidate in the database directly.

Sadly, the `bench_consolidation` benchmark doesn't show any speed
increase (or decrease!). The `bench_consolidation_large` benchmark takes
so long to run you need to decrease the sample count, but at that count
it varies so wildly as to be useless. The benchmarks are also performed
locally, on a single instance cockroach running on the same machine,
perhaps not a good proxy for performance in the cloud where requests
have to traverse a network.

Thus, the justification for this commit is that we're pretty sure it's
better, but don't yet have the data to support that conclusion.