storage: only source operators should carry tokens
The token mechanism is there to notify a timely operator this producing
data as a source (i.e does not have an input frontier to react to) that
its output is no longer needed and therefore it should terminate by
immediately advancing its output frontier to the empty antichain.

Out of confusion (at least speaking about me) we were also installing
and constructing tokens and were using them for all sorts of operators
when it was not necessary, and it was in fact problematic.

Every dataflow must ideally terminate gracefully which in timely terms
means that all operators drop their capabilities, everything reaches the
empty antichain and timely cleans up the resources.

Therefore, we only need to have a token for operators that produce data
on their own and everything else must rely on the frontiers of its
inputs to make decisions. This is true for all operators including our
persist sink or external sinks.

Enforcing that dataflow terminate gracefully and always reach the empty
antichain means that the storaged process will *always* send a
FrontierUppers response back to the storage controller that is at the
empty antichain. This is the moment for anyone listening to the feedback
stream to cleanup any associated resources since this is the moment that
no more updates will arrive after.

We were previously eagerly cleaning up the state of various things when
the since frontier advanced to the empty antichain (by observing
AllowCompaction commands), but then the components were confused when a
stray FrontierUppers made it their way and panicked. This is fixed by
waiting for the final terminating frontier to arrive before cleaning up.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>