persist: introduce "critical" SinceHandles
In contrast to [crate::read::ReadHandle], which is time-leased, this
handle and its associated capability are not leased. A SinceHandle only
releases its since capability when [Self::expire] is called. Also unlike
`ReadHandle`, expire is not called on drop. This is less ergonomic, but
useful for "critical" since holds which must survive even lease
timeouts.

**IMPORTANT**: The above means that if a SinceHandle is registered and
then lost, the shard's since will be permanently "stuck", forever
preventing logical compaction. Users are advised to durably record the
intended [CriticalReaderId] _before_ registering a SinceHandle (in case
the process crashes at the wrong time).

Touches #15511