persist: run background heartbeat tasks on CpuHeavyRuntime
(They're not actually cpu heavy.) It's critical that these heartbeats
run on time. Given that any instances of blocking work accidentally
running on the common Runtime could prevent that from happening, run
them on the one that happens to be entirely under persist's control. (We
should definitely fix any instances of blocking work on the common
runtime, this is just persist being defensive.)

Touches #15209