*: introduce halting as a softer way to panic
Introduce the concept of "halting", which a middle ground between
gracefully exiting a process and panicking. A halt indicates an error
has occurred that is severe enough to require shutting down the process,
but not severe enough to be considered a crash. Halts are not intended
to be reported to error tracking tools (e.g., Sentry).

There are two common classes of errors that trigger a halt:

  * An error that indicates a new process has taken over (e.g., an error
    indicating that the process's leader epoch has expired).

  * A retriable error where the retry logic is too difficult to write
    correctly. Sometimes the easiest way to retry an error is to tear down
    the process and allow a new process to start from scratch.

Halts are initiated via a new `halt!` macro in the `ore` crate.  Several
expected panics in the `persist`, `compute`, `storage`, and `adapter`
crates are downgraded to `halt`s, so that they don't unnecessarily
accumulate in Sentry.

Fix #15771.
Fix #15772.