zippy,platform-checks: Clean up Process Orchestrator metadata on restart
In order to work around #15725, clean up the Progress Orchestrator
information on PIDs and TCP ports on restart. As the tests are running
in containers, the processes that may have held those PIDs and ports
are now gone anyway when environmentd restarts, as they lived in
the same container.

Relates to ##15725, #15155