Coord: Implement timeout framework for clients
This commits adds the functionality to register, remove, and execute
client timeouts in the Coordinator.

This is very roughly inspired by the PostgreSQL timeout framework:
https://github.com/postgres/postgres/blob/fbfe6910eca0d6a61cbcdd27cdd5a8d9de6477fa/src/include/utils/timeout.h
https://github.com/postgres/postgres/blob/d37776e451e3481d7364137a0b159102bb05992a/src/backend/utils/misc/timeout.c

Currently we only use this framework for the
idle_in_transaction_session_timeout
(https://www.postgresql.org/docs/current/runtime-config-client.html).
However, it was designed to allow additional timeouts to be added in
the future.

Works towards resolving #13633