computed,storaged,timely-util: halt, not panic, on TCP errors
This commit is an extension to c0be2f86f, which introduced "halting" as
a softer way of panicking, to TCP errors in timely. TCP errors are
expected when a peer process in the timely cluster crashes. By
downgrading TCP errors from panics to halts, we'll focus attention on
the process that initially crashed, rather than distracting folks with
backtraces in the processes whose fate was shared.