storage: don't print backtrace when retrying indeterminate table write
This log line about retrying indeterminate table writes was including a
backtrace, which made it look like `environmentd` had crashed when in
fact it was happily retrying an error. Switch from the `Debug`
implementation to the alternate `Display` implementation for the error,
which will include the error's cause chain but not a backtrace.