Coord: Separate capability of storage and compute
Previously the Coordinator stored the read capability of all objects in
a map whose key was Global ID. For most objects this works fine since
most Global IDs refer to a single object. However for some objects,
such as materialized views and certain introspection sources, a COMPUTE
object and a STORAGE object share the same Global ID. This caused the
two objects to use the same value in the read capability map which led
to various inconsistencies.

This commit updates the Coordinator's state to store COMPUTE and
STORAGE read capabilities separately.

Fixes #15887