catalog: remove schema of unreferenced tables in PG sources
We previously retained table information for all tables in an upstream PG
source's publication, even if the source did not use those tables. In the
work of supporting unsupported types as text, it provided a superiror UX to
only track the information of referenced tables. This migration applies that
same logic to existing sources, and only retains the table information of
referenced tables.