adapter,sql[-parser]: allow referencing AWS PrivateLink connections
This commit adds the necessary plumbing to allow referencing AWS
PrivateLink connections in Kafka, PostgreSQL, and Confluent Schema
Registry connections.

Since this feature is highly unstable, gate it behind a new system limit
named `max_aws_privatelink_connections` that defaults to 0. This will
prevent users from creating this connection type when the release ships
on Wednesday, but we'll be able to opt environments in on a case-by-case
basis once we've performed our own internal testing.

This commit also consistently uses the spelling "AwsPrivatelink" and
"aws_privatelink", for consistency with how we've decided to spell this
in the external SQL syntax.

I wouldn't normally put so many unrelated changes into a single commit,
but we're under time pressure here.

Co-authored-by: Cara Haas <cara@materialize.com>
Co-authored-by: Alex Hunt <alex.hunt@materialize.com>
Co-authored-by: Joseph Koshakow <koshy44@gmail.com>