Clean-up of leaking `CollectionState` metadata in compute controller (#15862)
This commit ensures that collection metadata gets cleaned up from an Instance when all replicas in the current replica set of the Instance eventually report an empty frontier for the collection and the read frontier for the collection also becomes empty. The replica reporting is done by a FrontierUppers response, which is triggered by an AllowCompaction command to the empty frontier. In a replay of the compute command history, an AllowCompaction with the empty frontier renders the dataflow no longer live, so its collection metadata can be garbage collected as long as we know that all replicas processed the latter command. However, even if there are no replicas in a cluster, a collection may need to be still kept around, e.g., if it was created in response to DDL such as CREATE INDEX. So we can only garbage collect a collection when other components in the system manifest that it is safe to drop it by advancement of its read frontier to empty. 

The definition of current replica set here takes some care, as we may, e.g., explicitly change the replica set managed by the Instance by dropping or adding replicas. So the clean-up must leave the collection metadata of the Instance in a well-known state even as we interleave peeks, replica rehydration, replica addition / removal, and explicit collection addition / removal (e.g., by creating / dropping indexes). This commit focuses on necessary changes to the Instance implementation under the assumption that adapter-led recovery in the case of a crash of the compute controller correctly recreates only necessary collection metadata.