sql: Use checked recursion when planning set expressions
These calls potentially go quite deep, even with simple-seeming
queries; and the stack size of this frame is pretty large -- 62k for
me in debug, though only 4k in release -- so it doesn't take that much
to overflow here without these guards.

Also add a guard to plan_nested_query as that is mutually recursive
with many of the other planning functions.

Fixes #17075.