timely-util: ode to the polonius crab
This PR is using an insane crate to simplify and increase confidence on
the soundness of the implementation that was previously working around a
borrow checker limitation manually.

This crate puts some guardrails and ensures that only this particular
safe case is allowed to be expressed which in turns means that we won't
accidentally introduce unsoundness in the future as we're changing this
method.

Absolutely crazy.

h/t to Gus that showed me this crate

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>