[storage] Don't panic if someone else moves table upper (#15999)
The message in this panic is no longer true now that we persist sink status updates. The panic was there (presumably?) because we thought we would no longer make progress because there was a new leader. However, given that's not true, it should be sufficient to return an error.

It's also worth noting that, before this change, there are two different behaviors depending on whether we get an indeterminate error or not:

Read the upper from the writer, do a compare_and_append, get an Upper error because someone else advanced in the meantime, we return error from the function
Read the upper from the writer, do a compare_and_append, get an Indeterminate error, re-read the upper and notice that someone else has advanced in the meantime, we panic

Motivation
Fixes #15949