storage: in persist_source, hold lock only as long as we have to
The lock is serializing operator startup, which shows up as higher
latency for peaks with large-worker-count compute clusters.
Additionally, we were holding the lock for longer than necessary because
the mutex guard would only be dropped after creating the read handle.

Now, we lock, create the `PersistClient` and then release the lock,
meaning we do more things concurrently. The lock is _still_ serializing
startup, but not as much as before this change.

Touches #16090