storage: statically enforce definite errors
This PR changes the interface of `SourceReader` such that
implementations are required to provide a partition and offset whenever
they want to return an error. The errors are then reclocked just like
normal messages and emitted at the respective timestamps.

It's not clear if the existing source implementations emit definite
errors. This exercise forced me to come up with timestamps in some
places which indicates that they might not be definite. I marked the
relevant places in XXX comments so that we look into them.

This PR also sets the scene for the next step after native timestamp
reclocking which is to make the whole ingestion pipeline work with
native timestamps which makes it impossible to fabricate error
timestamps and diffs outside of the source readers since the timestamp
types become generics.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>