test: Increase max table limit in all slt tests
A test in our sqllogic git submodule,
test/sqllogictest/sqlite/test/select5.test, was failing because it was
trying to create more tables than the default max table limit. Since
this submodule is a fork and we try not to alter the tests, especially
with Materialize specific syntax, we couldn't increase the max table
limit in the test itself. Instead this commit increases the max table
limit for all sqllogic tests.