adapter: Prevent resource leaks in connection DDL
For certain connections, such as SSH and AWS PrivateLink, an external
resource was created before `catalog_transact` was run. If
`catalog_transact` failed, then those resources would be leaked
forever.

This commit updates the connection DDL so that the external resources
are only created if `catalog_transact` was successful.

Fixes #16072