persist_source: stop using DropSafeLeaseStream
We observed that the "LeaseStreamDrainer" task spawned by
`DropSafeLeaseStream::drop` sometimes never terminates because the
stream it is reading from is never exhausted, leading to a bunch of
tasks being left behind whenever a materialized view dataflow is
dropped. Since #15868, draining leased batches is not necessary anymore
to prevent panics, so just removing the `DropSafeLeaseStream` is fine.

This reverts commit 0739b020a6278c5ebf3da5230668be6363153422.