adapter: fix lifetime errors
There were two types of closures that were being passed to
`catalog_transact`. Closures that were returning some async work that
didn't need the passed `CatalogTxn` at all, and closures that were doing
synchronous work and needed the passed `CatalogTxn`.

Because in the latter case the code returned a `Future` that described
its synchronous work it was also forced to capture the `CatalogTxn`,
which was lifetimed, into the Future state, but that didn't pass the
borrow checker. In fact, I think it's impossible to express the lifetime
containts with the current vocabulary offered by rust.

However, we were overconstraining ourselves by staging *synchronous*
work into an async `Future`. The solution is to do all the synchronous
work that uses the `CatalogTxn` *before* returning the result. Since the
result itself doesn't carry any lifetimes the borrow checker is happy.

In this commit I also added a helper method for the very common case
that we want to perform a catalog transaction without performing any
additional actions.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>