persist: fix maelstrom flake
This was just a brain fart when updating maelstrom to try to maximally
exercise critical since handles in #15822. Before that PR, each copy of
Transactor had a ReadHandle and thus its own since capability. This
allowed it to know and assert that it could read at read_ts because it
only ever updated the since capability itself to things that were less
than read_ts. 15822 changed it so all Transactors shared a single
CriticalReaderId and thus a single since capability, so it's possible
for some other Transactor to move it past your read_ts. If we were able
to proceed with this read_ts, we'd just find out in compare_and_append
that it's out of date and we'd get a new read_ts that is above the
since_ts.

To fix, give each node a unique token (specifically, the maelstrom
node_id) for CaDS calls. When a CaDS detects that another node was the
most recent to update the since, forward the local read_ts to the
since_ts (i.e. set it to since_ts if the since_ts is larger).

Closes #16149