coord: Protect timestamp from crash loop (#16096)
Whenever a new Coordinator starts it resserves a new timestamp range in
the stash, larger than any previous timestamp range. This usually
results in the Coordinator resrving a range ahead of the system clock's
current time, because the range with the current time is owned by the
previous Coordinator. This will usually correct itself fairly quickly
because the Coordinator can advance it's timestamp much slower than
real time. However, if the Coordinator enters a crash loop, then the
the persisted timestamp can start to significantly exceed the system
clock as each new Coordinator reserves an even greater timestamp range.

This commit prevents a new Coordinator from starting with a timestamp
larger than a certain amount ahead of the system clock. Instead the
Coordinator sleeps and waits for the system clock to catch up. The
range was selected so that startup is not affected in the common case,
but in a crash loop or other error scenario no permanent damage is done
to the persisted timestamp.

Additionally, while running if the timestamp exceeds the system clock
by that same amount, we print an error.

Fixes #16007
Works towards resolving #14899