Downgrade k8s service watch errors
Watching k8s services sometimes produces errors. The ones we have
observed so far are:
 * "too old resource version"
 * "Connection refused"

In both cases, the errors seem to have been transient. The first ones
seem to be expected when watches are re-created, the second ones might
be caused by connection problems.

There is not much value in showing transient k8s errors in Sentry, so
this commit downgrades them to warnings.