adapter: Properly load indexes on system objects
Materialize allows users to create indexes on system objects to help
speed up debugging queries. Unless otherwise specified, these indexes
get created in the ambient database in the same schema as the
underlying object.

When loading user objects from the stash, we ignore objects from the
ambient schema. This caused Materialize to ignore certain indexes on
system objects when it restarted.

This commit updates the user object loading logic to not ignore objects
in the ambient database.

Fixes #16289