persist_sink: remove persist writer heartbeating
Moving to the "new style" async operators broke the activator-driven
heartbeating of the compute persist_sink. As it turns out, explicit
heardbeating of persist writers is not needed anymore, because persist
writers spawn their own tasks to do that now. So this commit removes the
heartbeating logic from persist_sink.