Limit the number of retries on retriable transaction errors
The design philosophy for sinks is that, on persistent errors, the sink
should eventually halt. This is especially important now that we track
the status of the sink, so errors are reported promptly.

Anyways, now we have a smallish retry limit on each of the two
operations. I've also converted these back to a more functional style,
since otherwise it'd be tedious to track the previous error across loop
iterations.