storage: add flow control to persist source
This PR adds the ability to flow control the output of `persist_source`.

Users of `persist_source` have the ability to apply flow control to the
output by providing a timely stream whose only purpose is to communicate
the frontier until which data must be emitted. When the frontier of that
flow control input is reached the source will stop producing data and it
will wait for the frontier of the flow control input to advance. When
the flow control frontier advances past the data frontier data emission
resumes. This effectively limits the number of timestamps that can be
in-flight at any given moment.

If no flow control is desired an empty stream whose frontier immediately
advances to the empty antichain can be used. An easy easy of creating
such stream is by using [`timely::dataflow::operators::generic::operator::empty`].

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>