storage: report native frontier when tracing
This PR fixes a bug where if tracing is enabled and the reclocking
operator is not initialized we might attempt to log its current upper as
an OffsetAntichain. However during initialization the source upper is
the empty antichain which cannot be represented as an OffsetAntichain.

In other words this code tried to be helpful by logging a trace and
instead panicked the process.

The bug only triggers when tracing is enabled which is how this got
through CI and extensive testing.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>