environmentd: periodically report environment summaries to Segment
This commit adds a telemetry loop, similar to what we had in the old
binary days, to environmentd.

Every hour, we report:

  * the current active sources, sinks, and subscribes as properties of
    the organization to Segment, via a group call.

  * the number of SELECT and SUBSCRIBE statements made in the last hour,
    as a "Environment Rolled Up" track event.

The idea is that tools that want the most up-to-date description of an
environment can look at the rolled up state in the `accounts` table,
while tools that want an hour-by-hour accounting of how active an
environment is can watch for "Environment Rolled Up" events.

Fix MaterializeInc/cloud#4734.