timely-util: automatically drain dropped handles
When writing async operators that are also tokened it's easy to
accidentally exit the future with data still pending on the inputs. This
can cause the dataflow to be permanently in memory and never dropped
because these inputs will hold back the frontier from ever advancing to
the empty antichain.

This change makes it so that when a handle gets dropped it registers
itself in a drainage queue that is managed by the async operator
machinery. When the logic future becomes ready (which means that the
user logic returned) the operator framework will automatically drain any
registered handles until timely eventually cleans it up.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>