adapter: Decrease bootstrap memory requirements (#16428)
Previously, when a caller wants to perform a write to a system table,
each write was represented by single `BuiltinTableUpdate` which was
converted to a single `PendingWriteTxn`. Each `PendingWriteTxn` is
fairly large (1,896 bytes), while each `BuiltinTableUpdate` was fairly
small (40 bytes). This generally caused a ballooning of memory
allocated for each system table write.

This was especially bad during bootstrapping which retracts all rows
in the system tables, and then re-inserts each row. This process could
occasionally require in excess of 2GB of memory.

This commit allows the caller to group a `Vec` of `BuiltinTableUpdate`
into a single `PendingWriteTxn`, significantly decreasing the size
needed.

Fixes #16404