ore,storaged,computed: introduce abstractions over Unix/TCP sockets
Add `SocketAddr`, `Listener`, and `Stream` types to the `mz_ore::netio`
module, which abstract over TCP sockets and Unix domain sockets. Then
teach storaged and computed to accept their listen addresses using the
new `SocketAddr` types, which allows them to bind to either TCP or Unix
domain sockets, as desired.

This is a key part step towards fixing the process orchestrator
(#15725), as it will allow multiple copies of Materialize to be run
concurrently without competing for access to the same ports.