Relax our behaviour on upsert semantics
This is a weird one!

The previous strict check here is a bit surprising:
- It can't actually prevent all invalid key collisions: it will notice
  if two records show up with a single key at a single timestamp, but it
  would be just as illegal if they showed up one timestamp apart.
- For debezium records we just publish all the records with a shared
  key, instead of doing anything fancy.

So even if it's Against the Rules to have multiple records with the same
key here, it doesn't seem worth trashing the process over.

There are a few ways one might choose to handle this. I've chosen to
just keep all the messages, to align the upsert behaviour more closely
with the debezium implementation, but other reasonable choices are
possible.

In general it would be nice to detect these errors, but I'm not sure one
can do it in the general case without keeping a ton of state, and I'm
not sure how valuable the non-general case is. (It certainly is capable
of catching _some_ errors, but seems quite delicate.)