*: use hakari to manage a workspace-hack crate
workspace-hack crates speed up builds when working across a workspace
with many crates by ensuring that all crates use the same features for
all transitive dependencies. For details, see the excellent hakari
documentation on the subject. [0]

This commit uses hakari to generate such a workspace-hack crate, and
teaches CI to enforce that it is kept up to date.

[0]: https://docs.rs/cargo-hakari/latest/cargo_hakari/about/index.html