compute: improve ReduceCollation unwrap error
This commit adds more detail to the panic message shown when
ReduceCollation fails to find all expected inputs. In particular, it
adds the offending key to the message.

We see ReduceCollation crash in production, but so far we don't have
enough data to diagnose the cause. The hope is that the improved error
message can help us.