adapter: Implement MVP for real time recency
This commit implements an MVP for real time recency on the adapter
side, gated behind unsafe mode.

Currently, the Coordinator waits for a recent timestamp in the main
coord loop, which can potentially block the Coordinator for an
unbounded amount of time. Before unsafe mode is removed, this waiting
will have to be removed from the main Coordinator loop.

Works towards resolving #11531