adapter: Fix serializable read-then-writes
Previously, how we picked a read timestamp for read-then-write queries
depended on what transaction isolation level that we were in. The
result of that read was then used to retract values from the underlying
table. If the read was performed in Serializable mode, then we may have
accidently read stale data and retracted stale data. This would
permanently corrupt the table making it unreadable.

This commit changes the timestamp selection process so that we always
pick a timestamp with the freshest data for read-then-write queries.

Fixes #16507