clusterd: unify storaged and computed
Unify storaged and computed into a single binary named "clusterd". Both
the storage controller and compute controller now create "clusterd"
services, and simply ignore the half that they do not need.

The ultimate goal is to allow scheduling multiple sources, sinks, and
indexes onto the same physical compute resources (i.e., a "cluster.")
This PR is the next step on that path. See [0] for details.

[0]: https://www.notion.so/materialize/Cluster-unification-472efd2e9b8d4286bacc91b84d6b6f71