storage: add WAL for shard truncation
If we intend to truncate a shard but crash before doing so,
persist does not have a mechanism to know that the shard should
not actually be in use.

To remedy this, set up a WAL that tracks the intent to truncate the
shard so that on reboot, the storage controller can do some bookkeeping
and truncate shards as needed.