ci/test: scope down the inputs to the cargo-test job
In 0275079 we switched to cargo-nextest, and in doing so accidentally
disabled running tests on PR builds (though they still ran on main).
77c6a9c re-enabled running tests on PR builds, but erred on the side of
caution, and always triggered the cargo test job if any file in the
repository changed. This causes a large performance regression for docs
only PRs, which we try to keep to less than one minute of CI time.

This commit uses a more precise filter for determining when to run the
cargo test job: whenever a Rust file changes, a Cargo.toml changes, or a
file in a "testdata" directory. There may be some edge cases that this
misses (changing a file that is `include!`d, perhaps), but it will catch
the vast majority of PRs that could cause a cargo test failure, while
restorign the CI speed of docs-only PRs.