test/upgrade: use a standalone CockroachDB instance
The legacy upgrade tests were previously using the CockroachDB instance
*inside* the `materialized` Docker image. That meant that when we
upgraded the version of CockroachDB inside the Docker image, we started
testing a CockroachDB version upgrade in addition to a Materialize
upgrade. Crucially, this doesn't work when we need to *downgrade* the
version of CockroachDB in the Materialize Docker image.

Switching to a standalone CockroachDB instance allows us to restart and
upgrade the Materialize version independently of CockroachDB.