repr: add `Row::unpack_owned` method
This commit adds a method to `Row` that allows the caller to unpack the
datums into an owned object so that the allocation and the work done to
figure out the datum boundaries can be reused.

This is done by using the `yoke` crate which is a way of defining
self-referencial data structures.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>