compute: configure DD's idle_merge_effort
This commit sets the `idle_merge_effort` configuration option of
differential dataflow, which controls how much work DD spends on
compacting arrangements. Anecdotically, the value 1000 provides a good
trade-of between CPU time spent compacting more eagerly and memory
saved by compacting earlier.

We might want to tune this value later, but it is clear that it should
not be `None`. Setting the option to `None` means that arrangement
compaction only happens in response to new updates arriving, so if a
dataflow receives no updates for a while, its arrangements may sit
around uncompacted, wasting precious memory.