timely-util: fixes for ProbeNotify
This commit fixes two issues with the new `ProbeNotify` mechanism:

 * It marks the `ProbeNotify` trade `pub`, so it can be used in other
   crates.
 * It makes the handle frontier default to the empty antichain instead
   of the minimal one.

The latter is needed to prevent probe handles that are not connected to
a stream, and therefore never have their handle frontiers advanced, from
holding back the global frontier.

Defaulting handle frontiers to the empty antichain means that a probe
that is not connected to any stream will always report the empty
frontier. The previous behavior was to report the minimal frontier
instead.