compute: intra-ts thinning for monotonic topk
This PR implements a pre-arrangement thinning of monotonic collections
that are on their way to a topk computation. This thinning has the
advantage of being able to be performed in a streaming fashion even for
single timestamps that might contain a lot of data.

With this change a monotonic collection flowing into a top 3 operator
whose snapshot is 10GB can be performed on machines with very little RAM
as we will incrementally discard records that cannot possible be in the
top 3 as rows flow in.