doc/developer: don't suggest binding port 8080 for Cockroach UI
CockroachDB runs its admin web UI on port 8080. The developer guide
previously suggested forwarding this port to the host machine to make
debugging Cockroach easier, but port 8080 is a common port and likely to
be in use by another tool.

Change the developer guide to instead suggest binding port
26258 on the host to port 8080 in the container. Port 26258 is much less
likely to be in use by another tool.

The admin UI port is not directly used by the Materialize development
environment, so there are no other references to this port that need to
be adjusted.