*: warn rather than deny lints
Using the deny level for lints prevents Materialize from compiling at
all with lint errors. This was not the intent of c513f09c7, and not how
bin/check previously worked. Instead, only CI should fail if there are
warnings; temporarily compiling with warnings is an important part of
the local development workflow.

This commit adjusts gen-lints to get the desired behavior. Lints are
installed with the warn level rather than the deny level, and then all
warnings are turned into denies in CI.