adapter: introduce the concept of a linked cluster
Part of the cluster unification project. See MaterializeInc/cloud#4929.
This is step 2 in the design document [0].

This commit introduces "linked clusters": clusters whose lifetimes are
tied to another object. Linked clusters preserve the soon-to-be legacy
interface for sources and sinks, where the size is specified on the
source/sink rather than the cluster replica.

You cannot managed linked clusters independently of the object they are
linked to. You cannot drop a linked cluster. You cannot create
additional replicas of a linked cluster. You cannot drop the one
automatically created replica of a linked cluster (named "linked").

When you drop a source or sink, Materialize automatically drops the
linked cluster. When you alter the size of a source or sink, Materialize
automatically drops the existing replica of the cluster and creates a
new replica of the new size.

Linked cluster operations are emitted to the `mz_audit_events` table.
This will allow the billing pipeline to be entirely in terms of compute
replica events; it will not need to separately consider source and sink
events.

Materialize exposes the links between cluster and sources/sinks in a new
`mz_internal.mz_cluster_links` system catalog table.

This commit includes a migration that will backfill linked clusters for
all existing sources and sinks on startup, including audit events for
the clusters' creation. This migration can be removed after it has been
included in one release, as going forward all new sources and sinks will
have linked clusters.

[0]: https://www.notion.so/materialize/Cluster-unification-472efd2e9b8d4286bacc91b84d6b6f71