compute: remove a panic from reduce implementation
The output type of SumUint64 is Numeric, which can handle negative
values, but the reduce implementation was panicking on negative values
by casting to Numeric via an unsigned type (i128 -> u128 -> Numeric).

This commit removes the panic by casting directly from i128 to Numeric,
as we're moving in the direction of making reduce less panicky (see
issue #8066).

This mitigates the crash we were seeing in #_incident-28, in which a
canceled query that used SumUint64 could result in the reduce operator
observing partial data and panicking due to the negative accumulation.

This approach is a bandaid, though, as SumUint16 and SumUint32 will
still panic in this situation, as their output type is Uint64, and we
have no choice but to panic in the reduce implementation until #8066 is
resolved.