storage: use critical since handle to hold back since of remap shard
Before, we were using only a leased reader for holding back the since of
the remap shard. This could lead to problematic cases because we
typically also have a second read handle (inside the subscription
handle) that also advances the since. If, on restart, we would expire
both handles we could run into situations where the shard-global since
advanced to the since of the "subscribe handle", which would put the
remap shard into a state from which we can't restart anymore.

It seems very natural that the remap operator, as the only user of the
remap shard, has a critical handle on it's since make sure that it stays
where it needs to stay.

Fixes #16421