mz: Add the ability to store app passwords within the local keychain for more secure storage (#16862)
* mz: move app password behind vault abstraction

Adds a new vault abstraction for storing app passwords. Vaults
return tokens that can be used for later password retrival. This
commit contains a single vault type - Inline - which stores the app
password inline within the configuration file for the CLI.

This commit is primarily for threading the vaults and tokens
through the code base. It does not modify any user facing functionality.

The vault is intended to be later extended to provide other variants
such as keychains, 1password, etc.

* mz: store app password in local keychain on macos

Adds keychain support when compiled for MacOS