*: remove some unnecessary mz_persist_types::Codec impls
The only types that need to impl Codec are ones uses as the key or value
of a persist shard. It's been fine so far for other things to impl
Codec, but when we add structure to persist, a Codec impl will be more
involved. So, remove the unnecessary impls to avoid the complexity.