storage,compute: simplify thread/operator activation by relying on async activation
Before, we were handing around Activators, so that we can activate
either the receiving thread or operator when sending compute/storage
commands.

Now, we use a tokio channel and use an async operator as the command
pump. This gives us activation of the timely operator "for free" and we
can get rid of the manual activators.

This requires that storage also adopt an internal control
dataflow/fabric for pumping commands, which in turns makes the
reconciliation logic a tiny smidge more complicated. We want to move in
this direction anyways, for supporting the distribution of internal
storage commands.

The motivation for this refactoring is that I want to add a way of
distributing replica-internal (internal to the timely cluster) commands,
for example for suspending and restarting a dataflow. One approach for
this would be to distribute commands using a timely-internal fabric,
similar to what this PR does. The subtle part is that we have to make
sure that the interleaving of internal and controller-issued commands is
the same on all workers, so they have to go through the same command
dataflow. And it seemed easier to work with tokio/async channels
compared to using "regular" channels and manually passing around
Activators.

We can also do that change (the internal command dataflow) without this
refactoring, it will just require more code around handling Activators
and the operator code that reads from both channels (internal and
external commands) will likely be more complicated.