zippy: Use a separate storage clusterd
Spawn a separate clusterd for sources and sinks so that Zippy
can kill and start it separately.

This provides additional event interleavings not possible before,
as the following events can happen while the storaged cluster is down:
- environmentd restarts
- Kafka events and table inserts
- materialized views creation

Bring the execution time of the various Zippy scenarios in Nightly
to ~20-30 min each.