sql: fix aliasing of output of NATURAL joins
Given the following two tables:

    CREATE TABLE t1 (a int);
    CREATE TABLE t2 (a int);

Materialize was incorrectly producing an error with this query:

    SELECT y.a FROM (t1 NATURAL JOIN t2) AS y;
    ERROR:  table reference "y" is ambiguous

The above query is meant to be equivalent to the following query,
which Materialize does correctly handle:

    SELECT a FROM (t1 NATURAL JOIN t2);

The bug was the result of the special handling of output join columns in
NATURAL/USING joins. When the output is unaliased, the input columns
remain available (in this example, as `t1.a` and `t2.a`). But when the
output is aliased, those input columns become unnamable. Materialize was
handling this correctly for `*` expansion, but not alias application.

Fix #16920.