reclock: read holds and compaction commutativity
This commit implements two changes in the ReclockFollower.

The first and simpler of the two is per-handle read holds so that each
user of a `ReclockFollower` can independently compact the in-memory
trace while having the trace be compaced that the minimum of all the
handles.

The second one is removing explicit tracking of the source since
frontier which was used to generate the `NotBeyondSince` error when
reclocking. It turns out that this breaks a useful property of the
reclock operator which is to commute with compaction. Specifically the
property that breaks is that:

```
compact(reclock(collection, trace)) == reclock(collection, compact(trace))`
```

Beyond just fixing the property for purity's sake, there is also a
practical aspect which is that it is incorrect to even attempt to filter
source timestamps using the calculated `source_since` frontier. Imagine
that we are reclocking this partial order:

```
  ,B---...
 /
A
 \
  `C---...
```

With the remap trace being:

```
(A, 0, +1)
(A, 1, -1)
(B, 1, +1)
(C, 1, +1)
... more bindings
```

Now suppose we compact the remap trace to `{1}`, which results in:

```
(B, 1, +1)
(C, 1, +1)
... more bindings
```

And we would naturally calculate the `source_since` to be `{B, C}`. If
we use this frontier to reject reclock requests that are not beyond it,
we have lost the ability to reclock events that happened at `A`!. And
there is no other timestamp that events at `A` could have because
advancing `A` by `{B, C}` yields just `A`. In other words, the code was
incorrectly assuming that after compaction all timestamps are beyond the
compaction frontier, which is not true.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>