adapter: change audit log timestamp to use oracle
Use the oracle timestamp instead of system clock for the audit events
table. The system clock was used because (at the time) the oracle ts
didn't have a firm bound on being off from the system clock, so using it
to determine when a thing happened was not safe. For example, we once
had a bug where the oracle ts was days ahead due to a fast loop. We've
improved the logic enough that we have enough confidence that won't
happen again (due to the use of write groups) that we'd now prefer to
have the guarantee of always goes up, and believe that the audit log
timestamps will be within a 10s margin of error.