test: remove idle_arrangement_merge_effort.td
This test was flaky because it can happen that the `INSERT` that is
supposed to trigger the consolidation of the previous insert +
retraction arrives at the replica before the index's since has advanced
beyond the time of the retraction. In that case, DD cannot merge the
retraction into the previous insert, as it has to maintain the
historical detail, and no consolidation is triggered.

Setting a logical compaction window of 1ms was supposed to avoid this
scenario, but obviously it is not enough. The only way I can think of to
ensure that the since has advanced far enough before sending the final
`INSERT` is to sleep, which does not seem like a robust solution either.