compute,adapter: generalize config update interface
To prepare for the addition of more compute configuration parameters,
this commit generalized the `ComputeController`'s config update
interface. Rather than taking only a single `max_result_size` parameter,
the compute controller now expects a `BTreeSet<ComputeParameter>`,
enabling other parameters to be passed as well.

The new `update_configuration` method is defined on the compute
controller, rather than the instance, and applies configuration updates
to all instances. This simplifies things for adapter. It also makes it
impossible to push configuration that is specific to a single instance,
but we don't need this functionality currently.

This commit also adjusts adapter code to deal with the changed
`ComputeController` API and also prepares it for the introduction of
more compute configuration parameters.