*: remove Stash trait in favor of a single impl
We only use the Memory impl for a few tests that can now safely depend
on cockroach existing.

We used the Cache impl for our production usages. That impl was added to
be a quick hack to speed up operations where we didn't know what was
going on. Since then we've greatly improved the underlying postgres perf
and added tracing so we have complete understanding of which statements
are executed. Furthermore, the Cache impl is only saving a small amount
of time by caching sinces and uppers, which are one of many things we
do.

Upcoming we expect to add a txn API for the stash to further reduce the
total number of SQL statements. This would be unimplementable by the
cache (thus not providing its caching speedup), and thus only result in
more typing for us.

We can instead simplify stash callers because we now only need the
single implementation.