compute: Move compute metrics to mz-compute crate
These metrics are solely used on the computed server side. They were
originally placed in the compute-client crate to re-use them in the
controller. However, they have to be different on the controller side,
as they need a replica label. This commit moves the metrics
to the compute crate where they are used.