adapter: Pull RTR off the main coord thread
Previously, when the Coordinator would ask the Controller for a real
time recency (rtr) timestamp, the Coordinator would wait for that
timestamp on the main Coordinator thread. This has negative performance
implications because it blocks the Coordinator from doing anything
else. This commit modifies the waiting so that it happens off the main
thread.

Additionally, a macro was added to help with sequence methods that
consume the client transmitter.

Fixes #16651