storage: produce definite pg snapshot
This PR is fixing a long standing issue with how postgres sources were
producing their initial snapshot. The problem was that because the
snapshot was being produced inside the same transaction that created the
replication slot, in order for the LSN of the slot and the snapshot to
match up, we only had one chance to produce the snapshot. Usually this
wasn't a problem because of a set of fragile conditions:

* The snapshot is transactionally committed to persist shards
* We always restarted at the upper frontier of the output

These two taken together meant that we could be smart and only produce
the snapshot when we were requested to ingest from LSN 0, but anything
non-zero implied that the snapshot had already been committed and so we
could skip it, saving ourselves from the difficult task of reproducing
it at the exact LSN of the slot.

It turns out that multi-output sources break the first condition as it's
possible that only a subset of subsource to commit the snapshots in
which case we'll have to recreate the snapshot on restart. The result is
a pretty serious correctness issue
https://github.com/MaterializeInc/materialize/issues/16865

This PRs fixes the problem by always producing the snapshot (when the
snapshot is needed of course) at the same LSN. The issue that the naive
approach has is that a database snapshot happens at an LSN that we don't
control, and it will almost always be at an LSN greater than the LSN of
the replication slot. The key observation is that just like a
differential collection can start from a snapshot and by adding diffs be
rolled forward in time we can also subtract updates and rewind it in
time. And it just so happens that we have the updates we need to
subtract in the replication slot!

At a high level the postgres source does the following:

1. Create the replication slot if it doesn't already exist and note down
   its `slot_lsn`
2. Take a snapshot in a separate transaction which will happen at some
   `snapshot_lsn` but emit all its updates at timestamp `slot_lsn`
3. If `slot_lsn < snapshot_lsn` start streaming the replication slot and
   for each update:
    * If the update's lsn is `<= snapshot_lsn` emit it at `slot_lsn`
      with negated diff
    * Once the stream goes beyond `snapshot_lsn ` stop rewinding and
      mark the snapshot as complete
4. Start the normal replication stream

In order to do the above a few changes needed to be done to the
structure of the postgres source:

1. The `produce_snapshot` and `produce_replication` methods are now
   freestanding functions that only get a pg client and give back an
   async stream with the updates without altering any state. This make
   it easy to instantiate the stream multiple times when needed.
2. Getting in and out of the replication stream to consume potential lag
   has been made a lot more aggressive and only based on time.

Finally, there was a bug with our `tokio-postgres` fork where dropping a
replication stream failed to put the connection in query mode. This is
something that has been fixed in the `rust-postgres` PR that is still
pending upstream (). For this reason I updated our fork to contain that
PR as-is and the replication specific code has been moved to
`mz_postgres_util`. This is the reason why this PR is so big.