Avoid multiple InspectBatch probing operators
Attaching a probe to a stream could only update a single handle. In many
cases, this meant that we would attach multiple InspectBatch operators to
a dataflow's output. This change alters this to only use a single operator
for a collection of probes.

Using the InspectBatch operator is not ideal because it forces the system
to create a complete copy of the input data, which is then immediately
discarded because we're only interested in progress statements. The fix
at hands reduces the number of InspectBatch operators, but does not avoid
the unnecessary copying of data.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>