sql: permit subqueries in SET clause of UPDATE
These were disallowed in a conservative approach to handling
ReadThenWritePlans. However, in the intervening time, we added
rigorous safeguards to ensure that the "read" portion of these
plans do not reference any potentially consistency-violating
objects (#2b3f712dd5), so this approach is no longer necessary.

Additionally, this constraint prevents using parameters in update
statements' set clauses, which is really annoying.