compute-client: support replica-targeted subscribes
This commit adds support for replica-targeted subscribes in the compute
controller.

Like replica-targeted peeks, we handle replica-targeted subscribes in
the compute controller by filtering the responses we receive.

Unlike replica-targeted peeks, passing the target replica for a
subscribe to the compute controller is slightly awkward because
subscribes are not created through a dedicated method but through
`create_dataflows`. Adding a `target_replica` argument to this method
seems like it would be confusing, since most dataflows don't export
subscribes and therefore have no use for this argument. So instead
replica-targeting for subscribes is configured by a separate method
`set_subscribe_target_replica`.