storage: in workers, add internal control dataflow
This is in preparation for adding internal control messages that can
restart a dataflow from within the workers, without involving the
controllers.

For that, we need to make sure that commands that cause dataflows to be
rendered are processed in the same consistent order across all workers
because timely requires this. This was previously ensured because each
worker was only working off a stream of commands from the controller,
and those commands where assumed to be in the same order across workers.

With the introduction of internal commands, which can also cause
dataflows to be shut down and (re-)rendered, we need to make sure we get
a consistent sequence across all of these.

To achieve this, we introduce internal commands and make sure that only
internal commands can cause dataflows to be rendered. These internal
commands go through a timely `Sequencer`, which ensures all workers
receive the same sequence of commands. External commands (from the
controller) now cause internal commands to be broadcast (by only one
worker), to cause dataflows to be rendered.