sql: Give QueryContext an ID generator that starts after CTE IDs
CTE LocalId generation happens in names::resolve(), without an ID
generator, before we get to planning a query.  Ideally, it would be
aware of other potential IDs in a query, and use an ID generator.

The approach we use here, to permit plan to introduce new Let
bindings, might seem inelegant, but the alternative, of trying to have
names::resolve take a query context or return that information, seems
like a much messier change: not all calls have a query context around,
and there are a bunch of them.

XXX I just realized we can plan_ctes recursively in nested queries and
this won't work.