persist: reduce S3 read timeout
It appears to be a common occurrence that reads from S3 time out the on
the first try but immediately succeed on the second try. The previous S3
read timeout was 60 seconds, which lead to cases where reads from
persist sometimes took a minute longer than usual, which in turn would
become very noticable in one-off queries reading from a storage source.

This commit attempts to reduce the impact to query times by aggressively
reducing the S3 read timeout. The new timeout is 4s, which would still
introduce a noticable delay, especially to otherwise fast queries, but
much less than the previous timeout.

4s seems like a safe timeout, since our persist clients are expected to
be located physically close to the S3 endpoints. We might even be able
to go lower than that after some testing.