Remove Diff parameter from persist's constructor methods
...and nowhere else, yet.

The thesis of this change is that we seem to roughly agree that we don't
need to be generic here: all known users are happy with this diff type,
there are compatibility arguments for why all users _should_ use the
same diff type, and it may make certain aspects of the implementation
simpler. (For example, not needing to have access to the semigroup to do
compaction.)

This doesn't actually drive through the refactoring all the way; it just
prevents users from using non-i64 diffs, but struct type params and
other usage stays the same. The next step would be to push the
de-D-ification of persist deeper into the implementation.