adapter,sql: support creating sources and sinks in clusters
Part of the cluster unification epic (MaterializeInc/cloud#4929).

This commit adds the `IN CLUSTER` clause to the `CREATE SOURCE` and
`CREATE SINK` statements to allow creating a source or sink in an
existing cluster. There are a few limitations:

  * Any number of sources or sinks can be created in the same cluster,
    but a given cluster can only contain sources/sinks or
    indexes/materialized views.

  * Clusters containing sources or sinks can have at most one replica.

We plan to lift these limitations in the future. See the cluster
unification design document (linked in the epic) for details.