adapter: separate cluster [replica] ID allocation from creation
Part of the cluster unification epic (MaterializeInc/cloud#4929).

Adjust the API for creating clusters and cluster replicas to allow
allocating the ID for the to-be-created cluster or cluster replica
outside of the transaction that creates the cluster or cluster replica.
This matches what we do for other catalog entities.

A future commit will use this ability to create sources and sinks that
refer to their linked cluster in single transaction.