Migration to remove legacy orchestrator services
This commit adds a special migration that removes all orchestrator
services in the 'compute' and 'storage' namespaces when environmentd
starts up.

In version 0.41.0, the controller only knows about the 'cluster'
namespace, so services that were created in other namespaces previously
leak. The purpose of this migration is to clean up those leaked
services.