compute-client: remove remove_replica_state method
We originally introduced this method to be able to execute a subset of
`remove_replica` that does not perform orchestrator actions. Now that
orchestrator actions are performed at a higher level, `remove_replica`
is just a thin wrapper around `remove_replica_state`, so we can inline
the latter into the former and avoid confusion about why the latter even
exists.