storage,compute: clarify `create_instance`/`drop_instance` naming
Before, the name was implying that the callee was doing the actual work
while in fact it is just being notified of things that already happened.

Now, prefix these methods with `notify_` to convey that fact.

NOTE: I can thread this all the way to `controller/clusters.rs` if we
like the change.