ore: add Hash collection wrappers
This commit adds wrappers for the stdlib collection types `HashMap` and
`HashSet` to `mz_ore`, the purpose of whichs is to provide a means to use
the Hash collections safely in contexts that rely on stable iteration
order. This safety is achieved by only re-exporting those methods that
don't expose the iteration order.

The new wrappers are, obviously, more limited than the stdlib types they
wrap, as they don't support iteration over the collection contents. I
expect that we might have to extend them in the future, e.g., by
providing opt-in, denied-by-default ways to perform iteration.