sql: remove uses of Hash collections
This commit removes all uses of `HashMap` and `HashSet` in `mz_sql`, by
replacing them with their BTree equivalents.

This requires deriving `PartialOrd` and `Ord` for a bunch of structs
that didn't need these impls before. On the plus side, we should be able
to remove many of the `Hash` impls once the Hash -> BTree transition is
complete.