adapter: Add deterministic audit event timestamps
This commit adds a toggle that allows users specify what timestamp will
be used for audit events. When the toggle is off, the system will
choose timestamps normally. This toggle is only available in unsafe
mode and is meant for testing purposes.

This commit also adds a toggle for all system variables that will make
them only available in unsafe mode.

Resolves #17331