use ConsolidateBuffer in explode_one
This commit changes the implementation of explode_one in timely-util to employ a ConsolidateBuffer. This change gains us the capability to consolidate across input batches. For the change to work in conjunction with an implementation based on a unary operator, a few adjustments to the typing signature of ConsolidateBuffer needed to be carried out.