ci: Add journalctl output to the buildkite artifacts
Buildkite hosts run an AMI image derived from Amazon Linux, which
in turn derives from Red Hat.

Turns out that on such systems the actions of the OOM killer are
not reflected in /var/log/dmesg. That file is only written to
during kernel bootstrap. The journalctl utility must be used
to capture any messages about the OOM killer.

The BuildkiteAdditionalSudoPermissions setting of the Buildkite
CloudFormation template is used to allow the buildkite-agent
to run journalctl using sudo.