ConsolidateBuffer improvements
This commit adds a give_iterator function to ConsolidateBuffer as well as evolves its logic on flush to take into account that give_vec does not guarantee that a vector of zero capacity will be returned after a call. Additionally, the commit adds documentation to give regarding why the buffer capacity there is not overshot. Given the improvements in ConsolidateBuffer, the explode_one usage is refactored to make it more idiomatic.