persist: push the Mutex into PersistClientCache
This is immediately motivated by #16797, which hooks up dynamic
LaunchDarkly configurations to PersistConfig. To do this, we need a
`&PersistConfig`, but it's currently passed around locked up inside the
Mutex around PersistClientCache.

I could plumb a copy of PersistConfig to the necessary places instead,
but we've discussed it being desirable to push the Mutex inside
PersistClientCache anyway. Specifically to systematically address a
category of bugs where the Mutex guard is accidentally held longer than
intended.