persist: remove K, V, D bounds from State
This is preparation for #17317, where we need access to State with
unknown (and potentially mixed) K, V. The T param has to stay (for the
foreseeable future) because the various frontiers in State are expressed
in terms of it.

This is accomplished by moving the codec validation later. Before, it
was done at the time we decode State from a proto, which was a nice
single codepath to make sure it was always done. Now, proto decode
returns a newtype wrapper around State that marks it as not yet having
codecs validated. This allows the type system to enforce that we
validate it in all the necessary places.