storage: reclock: take batches into account when downgrading
This issue appeared after I merged #17452 yesterday. Previous to that PR
we would never hold pending work in our operator state that was not
beyond the source upper frontier, and therefore it was correct to only
consult the frontier of the timestamper and the frontier of the source
handle to calculate what kind of capabilities we need to keep around.

However, after that PR it is possible that we keep around batches for a
little bit longer, after the source upper frontier has moved past them
and so it is required that we take these held onto times into account
when we calculate what is the frontier that is safe to downgrade to.

Fixes #17484

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>