storage: move methods into the `SourceConnection` trait
This PR is mostly code movement to capture the fact that each source
connection implements a common interface that is then packaged up in a
`GenericSourceConnection` enum that simply dispatches to the individual
connections.

By having a common trait that describes a connection we can now define
the `IngestionDescription` as a generic type. I've kept the default
value for the generic parameter to be `GenericSourceConnection` so that
most uses are unaffected.

This change is setting the stage and enables expressing the _entire_
ingestion dataflow rendering through a function that is generic over the
connection. We already have functions that are generic over the
connection, but only for the first step of ingestion. The return types
of these function are uniform so that we can continue with decoding and
envelope processing in a non connection generic manner.

We have faced the same problem in the past where depending on the source
we render we might get a different stream type. A postgres source gives
out a stream of Rows that can't go through decoding (the types don't
match) whereas Kafka gives out bytes which can be decoded. In the past
we solved this non-uniformity by introducing the `SourceType` enum which
can gather all the known types (byte delimited, byte stream, row source)
in one type, and then we later match again against it.

However if we want to push native timestamps even further we will want
to render a source and maintain its original timestamp in the stream,
which means that each source now will have a unique stream type because
somewhere in there there will be the `SourceReader::Time` type.

In other words, we currently have a `match connection { .. }` in
`create_raw_source` but we need to lift it upwards, probably in
`render_source`.

Hopefully all the above will make more sense once the code implementing
these ideas lands.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>