compute-client: fix partitioned subscribe tracking
The subscribe tracking in `PartitionedComputeState` immediately emits a
`DroppedAt` response when it sees one from one of the shards. It failed
to take into account that after one shard has sent a `DroppedAt`
response others might still send `Batch` responses. This made it
possible that `PartitionedComputeState` could emit a `Batch` response
after it had already emitted a `DroppedAt` response, which is not
allowed according to the compute protocol and confuses the compute
controller.

This commit fixes that bug by ensuring the `PartitionedComputeState`
does not emit `Batch` responses for subscribes that have been dropped
already.