sql: Add broken implementation for column disambiguation
This commit adds a broken implementation to re-write view definitions
so that they don't contain potentially ambiguous expressions.
Potentially ambiguous expressions are expressions whose meaning can
change if columns are added to underlying objects. These expressions
include:
  - Non-qualified column references.
  - `*` (wildcard).
  - `NATURAL JOIN`.
This is a necessary step to allow us to add columns to a table without
introducing ambiguity into a query. For example imagine the following
tables:
    `CREATE TABLE t1 (a INT)`
    `CREATE TABLE t2 (b INT)`
and the following view:
    `CREATE VIEW v AS SELECT a, b FROM t1, t2`
Here `a` is unambiguous because only `t1` contains a column `a`.
However, if we add a column `a` to `t2`, then `a` become ambiguous in
the view definition.

The approach taken is as follows. During planning we collect
all metadata needed to re-write the statement. Then after planning we
go back and re-write the statement with this information. Finally, we
convert the statement to text and save it in the catalog.

One unfortunate consequence is that all function re-writing and
de-sugaring is now saved in the view definition. This can lead to some
ugly looking view definitions.

As indicated previously, this implementation is broken, so it's
completely hidden behind a feature flag. We decided to merge this
because priorities are changing, and we don't want to lose the work
that went into it.

Part of #