sql: Fully qualify all column names in views
This commit updates SQL planning so that all queries are re-written
with fully qualifies columns. This is a necessary step to allow us to
add columns to a table without introducing ambiguity into a query. For
example imagine the following table:
    `CREATE TABLE t1 (a INT)`
    `CREATE TABLE t2 (b INT)`
and the following view:
    `CREATE VIEW v AS SELECT a, b FROM t1, t2`
Here `a` is unambiguous because only `t1` contains a column `a`.
However, if we add a column `a` to `t2`, then `a` become ambiguous in
the view definition.

The approach taken is as follows. During planning we collect
a mapping that contains the fully qualified name of each column
reference. Then after planning we go back and re-write the statement
with this information. Finally, we convert the statement to text and
save it in the catalog.

One unfortunate consequence is that all function re-writing and
de-sugaring is now saved in the view definition. This can lead to some
ugly looking view definitions.

Part of #