adapter: correctly compute current_schema
This function in postgres takes the search_path session var (which can
be a list in postgres but not yet in mz) and returns the list of schemas
that exist with those names. We were previously passing the session var
through without checking for existence.

Test output here matches my testing in Postgres (less the mz_catalog
schema).