compute: add reconciliation metrics
This commit adds Prometheus metrics that provide visibility into how
well compute reconciliation works. These metrics track the number of
successful reconciliations (i.e. dataflows that could be reused) as well
as the number of unsuccessful reconciliations (dataflows that had to be
replaced by fresh ones). For unsuccessful reconciliations a label is
provided that records the reason why the old dataflow had to be
replaced.